+++ apps/web/config/routes.rb
@@ -5,6 +5,6 @@ get "/",                         to: "dashboard#index",      as: :root
 resources :errors,         only: [:index, :show], controller: "external_errors"
 resources :reservations,   only: [:index]
 resources :suppliers,      only: [:show] do
-  resources :hosts,        only: [:new, :create], controller: "hosts"
+  resources :hosts,        only: [:new, :create, :edit], controller: "hosts"
 end
 resources :sync_processes, only: [:index]